backend2/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic    